test: add testing utils and split test suits
