+++ packages/api-headless-cms/__tests__/manageAPI/snapshots/product.js
