--- packages/api-headless-cms/__tests__/tests/manage/index.js
@@ -1,20 +1,0 @@
-import { setupSchema as setupTestingSchema } from "@webiny/graphql/testing";
-import contentModelToSDL from "./contentModelToSDL";
-import schema from "./graphqlSchema";
-import resolvers from "./resolvers";
-import headlessPlugins from "../../../src/handler/plugins";
-
-export default ({ plugins }) => {
-    function setupSchema() {
-        return setupTestingSchema([
-            plugins,
-            headlessPlugins({ type: "manage", environment: "default" })
-        ]);
-    }
-
-    describe("MANAGE API", () => {
-        contentModelToSDL({ plugins });
-        schema({ setupSchema });
-        resolvers({ setupSchema });
-    });
-};
