--- packages/api-headless-cms/__tests__/tests/utils/id.js
@@ -1,13 +1,0 @@
-import { string } from "@commodo/fields";
-
-export default () => {
-    return string({
-        validation: value => {
-            if (value && !(typeof value === "string")) {
-                throw new Error(
-                    `A valid ID must be assigned to the "id" field (tried to assign "${value}").`
-                );
-            }
-        }
-    });
-};
