--- packages/api-headless-cms/__tests__/tests/utils/withId.js
@@ -1,8 +1,0 @@
-import { withFields } from "@commodo/fields";
-import id from "./id";
-
-export default () => {
-    return withFields({
-        id: id()
-    });
-};
