+++ packages/api-headless-cms/__tests__/utils/index.js
@@ -1,3 +1,6 @@
+import i18n from "@webiny/api-i18n/plugins/service";
+import mockI18NLocales from "../mocks/mockI18NLocales";
+
 import useDatabase from "./useDatabase";
 import useApolloHandler from "./useApolloHandler";
 import useCopyHandler from "./useCopyHandler";
@@ -6,11 +9,12 @@ import useSchema from "./useSchema";
 import applyMongoDb from "./mongodb";
 
 export const createUtils = (plugins = []) => {
-    return applyMongoDb(plugins, plugins => ({
-        useDatabase: useDatabase(plugins),
+    const utilsId = Date.now();
+    return applyMongoDb(utilsId, [i18n(), mockI18NLocales(), ...plugins], plugins => ({
+        useDatabase: useDatabase(utilsId, plugins),
         useApolloHandler: useApolloHandler(plugins),
         useCopyHandler: useCopyHandler(plugins),
         useContext: useContext(plugins),
-        useSchema: useSchema(plugins),
+        useSchema: useSchema(plugins)
     }));
 };
