+++ packages/api-headless-cms/__tests__/utils/useContext.js
@@ -1,5 +1,5 @@
 import { setupContext } from "@webiny/graphql/testing";
 
-export default plugins => async contextPlugins => {
-    return await setupContext([...plugins, ...contextPlugins]);
+export default plugins => async (contextPlugins = [], baseContext = {}) => {
+    return await setupContext([...plugins, ...contextPlugins], baseContext);
 };
